MapLibreDeckGLMap.tsx works now. (#16)
Use it like:
  <MapLibreDeckGLMap style={{ height:'1080px', width:'1920px' }} />

A defined height and width are required, but they can be relative
measurements: ex. 100%, 100vh.

Currently queries everything within the viewport of the map.
Each time there's a drag or zoom event, a new query is rendered
accordingly.

frontend/public/index.html was updated with the static scripts
and CSS style necessary for deck.gl and MapLibre.